requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
flask-limiter>=3.5.0
bcrypt>=4.0.0
groq>=0.9.0

# Google APIs
gspread>=6.0.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0

# Document generation
python-docx>=1.1.0

# Email
# (uses stdlib smtplib — no extra package needed)

# Cal.com / HTTP
requests>=2.31.0
